requirements.txt
aiohttp==3.8.5
bardapi==0.1.38
emoji==2.8.0
google==3.0.0
googlesearch_python==1.2.3
googletrans==3.1.0a0
gTTS==2.3.2
instaloader==4.10
keyboard==0.13.5
nltk==3.8.1
numpy==1.24.2
openai==0.27.4
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==9.3.0
Pillow==10.1.0
plyer==2.1.0
psutil==5.9.5
PyAudio==0.2.13
PyAutoGUI==0.9.54
pydub==0.25.1
pygame==2.5.0
PyGetWindow==0.0.9
pynput==1.7.6
pyperclip==1.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pywhatkit==5.4
Requests==2.31.0
selenium==4.14.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
torch==2.0.0
wikipedia==1.4.0
wolframalpha==5.0.0
bardapi==0.1.38
Eel==0.16.0
engine==0.6.0
googletrans==3.1.0a0
gTTS==2.3.2
keyboard==0.13.5
nltk==3.8.1
numpy==1.24.2
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==9.3.0
Pillow==10.1.0
plyer==2.1.0
PyAutoGUI==0.9.54
pygame==2.5.0
PyGetWindow==0.0.9
pynput==1.7.6
pyperclip==1.8.2
pyttsx3==2.90
selenium==4.14.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
torch==2.0.0
